requirements.txt
pip==25.1.1
fastapi
uvicorn
pydantic
httpx
python-dotenv
sqlalchemy
aiosqlite
fastapi-users[sqlalchemy]
jinja2
python-multipart
bcrypt
pyyaml
msal
openai
google-generativeai
anthropic
ollama
pypdf
python-docx
sentry-sdk[fastapi]
tiktoken

